Integrated credential checker into settings and setup flow
- adds new [projectId]/setup route
